requirements.txt
python-telegram-bot==21.10
google-genai==1.14.0
